Extract PolicyRuleSet integration tests from kernel.rs
Move the policy_rules_integration test module (5 tests, 180 lines) to
crates/portcullis/tests/kernel_policy.rs as a standalone integration
test using only public API. Reduces kernel.rs from 4,077 to 3,897 lines
and ratchets the ceiling from 4,080 to 3,900.
