+++ .line-ratchet.toml
@@ -20,6 +20,6 @@ path = "crates/nucleus-claude-hook/src/main.rs"
 
 [[files]]
 path = "crates/portcullis/src/kernel.rs"
-ceiling = 4080
+ceiling = 3900
 target = 2500
 step = 1
